+++ src/utils/config_parser/index.test.ts
@@ -1,4 +1,4 @@
-import { getOwnershipRules, parseOwnershipRules } from ".";
+import { getOwnershipRules } from ".";
 import { setConfigNotFound } from "../../../test/utils/config";
 
 describe("config parser tests", () => {
@@ -6,23 +6,5 @@ describe("config parser tests", () => {
     setConfigNotFound();
     const config = await getOwnershipRules(null);
     expect(config).toBeDefined();
-  });
-
-  test("parse basic config", () => {
-    expect(
-      parseOwnershipRules(
-        {
-          "ownership_rules": {
-            "directory_matching_rules": [
-              {
-                "name": "test_rule_0",
-                "path": "playground/{{username}}/*",
-              },
-            ],
-          },
-        },
-        null,
-      ),
-    ).toBeDefined();
   });
 });
