+++ src/utils/types/index.ts
@@ -14,6 +14,7 @@ export interface OwnershipRules {
    * is safe, this is disabled by default.
    */
   allowDotGitHub: boolean;
+  globalBlacklistedUsers: UserInfo[];
   directoryMatchingRules: DirectoryMatchingRule[];
 }
